requirements.txt
pyinstaller
python-dotenv

langchain
langchain-groq
langchain-google-genai